:art: split assertion tests to organize them better
